requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
mysql-connector-python>=8.1.0

# Web scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Task scheduling
schedule>=1.2.0
python-crontab>=2.7.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorlog>=6.7.0

# Optional: Notifications
discord-webhook>=1.3.0
